fixed many minor smells

+++ logicaldoc-gui/src/main/java/com/logicaldoc/gui/common/client/widgets/ImageCropper.java
@@ -1,16 +1,10 @@
 package com.logicaldoc.gui.common.client.widgets;
 
-import com.google.gwt.dom.client.ImageElement;
 import com.google.gwt.widgetideas.graphics.client.ImageLoader;
-import com.google.gwt.widgetideas.graphics.client.ImageLoader.CallBack;
 import com.smartgwt.client.types.DragAppearance;
 import com.smartgwt.client.types.Overflow;
 import com.smartgwt.client.widgets.Canvas;
 import com.smartgwt.client.widgets.Img;
-import com.smartgwt.client.widgets.events.MovedEvent;
-import com.smartgwt.client.widgets.events.MovedHandler;
-import com.smartgwt.client.widgets.events.ResizedEvent;
-import com.smartgwt.client.widgets.events.ResizedHandler;
 
 /**
  * Displays an image and allows the seletion of an area.
@@ -35,12 +29,12 @@ public class ImageCropper extends Canvas {
 	public ImageCropper(String imageUrl, int initialSelectionWidth, double selectionAspectRatio) {
 		setSelectionAspectRatio(selectionAspectRatio);
 
-		ImageLoader.loadImages(new String[] { imageUrl },  imageElements -> {
-				imageWidth = imageElements[0].getWidth();
-				imageHeight = imageElements[0].getHeight();
+		ImageLoader.loadImages(new String[] { imageUrl }, imageElements -> {
+			imageWidth = imageElements[0].getWidth();
+			imageHeight = imageElements[0].getHeight();
 
-				img.setHeight(getHeight());
-				img.setWidth((int) ((double) getHeight() * getImageAspectRatio()));
+			img.setHeight(getHeight());
+			img.setWidth((int) ((double) getHeight() * getImageAspectRatio()));
 		});
 
 		img = new Img(imageUrl);
@@ -56,30 +50,29 @@ public class ImageCropper extends Canvas {
 		selection.setLeft(100);
 		selection.setTop(100);
 		selection.setWidth(initialSelectionWidth);
-		selection.setHeight(
-				selectionAspectRatio > 0 ? (int) ((double) initialSelectionWidth / selectionAspectRatio) : 40);
+		selection.setHeight(selectionAspectRatio > 0 ? (int) (initialSelectionWidth / selectionAspectRatio) : 40);
 		selection.setOpacity(60);
-		selection.addResizedHandler( event -> {
-				if (maintainSelectionAspectRatio)
-					selection.setHeight((int) ((double) selection.getWidth() / selectionAspectRatio));
+		selection.addResizedHandler(event -> {
+			if (maintainSelectionAspectRatio)
+				selection.setHeight((int) ((double) selection.getWidth() / selectionAspectRatio));
 
-				if (selection.getLeft() + selection.getWidth() > img.getWidth())
-					selection.setWidth(img.getWidth() - selection.getLeft());
+			if (selection.getLeft() + selection.getWidth() > img.getWidth())
+				selection.setWidth(img.getWidth() - selection.getLeft());
 
-				if (selection.getTop() + selection.getHeight() > img.getHeight())
-					selection.setHeight(img.getHeight() - selection.getTop());
+			if (selection.getTop() + selection.getHeight() > img.getHeight())
+				selection.setHeight(img.getHeight() - selection.getTop());
 		});
 
 		selection.addMovedHandler(event -> {
-				if (selection.getLeft() + selection.getWidth() > img.getWidth())
-					selection.setLeft(img.getWidth() - selection.getWidth());
-				if (selection.getTop() + selection.getHeight() > img.getHeight())
-					selection.setTop(img.getHeight() - selection.getHeight());
+			if (selection.getLeft() + selection.getWidth() > img.getWidth())
+				selection.setLeft(img.getWidth() - selection.getWidth());
+			if (selection.getTop() + selection.getHeight() > img.getHeight())
+				selection.setTop(img.getHeight() - selection.getHeight());
 
-				if (selection.getLeft() < 0)
-					selection.setLeft(0);
-				if (selection.getTop() < 0)
-					selection.setTop(0);
+			if (selection.getLeft() < 0)
+				selection.setLeft(0);
+			if (selection.getTop() < 0)
+				selection.setTop(0);
 		});
 
 		addChild(img);
@@ -118,17 +111,18 @@ public class ImageCropper extends Canvas {
 
 	public void resize(int zoom) {
 		int newHeight = img.getHeight() + zoom;
-		int newWidth = (int) ((double) newHeight * getImageAspectRatio());
+
+		int newWidth = (int) (newHeight * getImageAspectRatio());
 
 		double proportionW = img.getWidth() / (double) selection.getWidth();
 		double proportionH = img.getHeight() / (double) selection.getHeight();
 		double proportionX = img.getWidth() / (double) selection.getLeft();
 		double proportionY = img.getHeight() / (double) selection.getTop();
 
-		int newSelectionHeight = (int) ((double) newHeight / proportionH);
-		int newSelectionWidth = (int) ((double) newWidth / proportionW);
-		int newSelectionTop = (int) ((double) newHeight / proportionY);
-		int newSelectionLeft = (int) ((double) newWidth / proportionX);
+		int newSelectionHeight = (int) (newHeight / proportionH);
+		int newSelectionWidth = (int) (newWidth / proportionW);
+		int newSelectionTop = (int) (newHeight / proportionY);
+		int newSelectionLeft = (int) (newWidth / proportionX);
 
 		if (newHeight > 100) {
 			img.setHeight(newHeight);
